autocxx
Module extern_rust

Tools to export Rust code to C++.

Attribute Macros

extern_rust_function

Declare that a given function is a Rust function which is to be exported to C++. This is used as an attribute macro on a Rust function, for instance:

extern_rust_type

Declare that this is a Rust type which is to be exported to C++. You can use this in two ways: