autocxx::extern_rust
Attribute Macro extern_rust_function

#[extern_rust_function]

Declare that a given function is a Rust function which is to be exported to C++. This is used as an attribute macro on a Rust function, for instance:

#[extern_rust_function]
pub fn call_me_from_cpp() { }

See extern_rust_type for details of how to find the generated declarations from C++. Attribute to state that a Rust function is to be exported to C++ in the extern "Rust" section of the generated cxx bindings.