autocxx

Macro extern_cpp_opaque_type

macro_rules! extern_cpp_opaque_type {
    ($($tt:tt)*) => { ... };
}

Indicates that a C++ type is not to be generated by autocxx in this case, but instead should refer to some pre-existing Rust type. Unlike extern_cpp_type!, there’s no need for the size and alignment of this type to be correct.

The syntax is: extern_cpp_opaque_type!("CppNameGoesHere", path::to::rust::type)

A directive to be included inside include_cpp - see include_cpp for general information.