autocxx::subclass::prelude

Attribute Macro is_subclass

#[is_subclass]

Deprecated - use subclass instead. Attribute to state that a Rust struct is a C++ subclass. This adds an additional field to the struct which autocxx uses to track a C++ instantiation of this Rust subclass.