autocxx::subclass
Trait CppSubclassDefault

pub trait CppSubclassDefault<CppPeer: CppSubclassCppPeer>: CppSubclass<CppPeer> + Default {
    // Required methods
    fn default_rust_owned() -> Rc<RefCell<Self>>;
    fn default_cpp_owned() -> UniquePtr<CppPeer> ;
}

Provides default constructors for subclasses which implement Default.

Required Methods

fn default_rust_owned() -> Rc<RefCell<Self>>

Create a Rust-owned instance of this subclass, initializing with default values. See CppSubclass for more details of the ownership models available.

fn default_cpp_owned() -> UniquePtr<CppPeer>

Create a C++-owned instance of this subclass, initializing with default values. See CppSubclass for more details of the ownership models available.

Object Safety

This trait is not object safe.