autocxx
Trait RValueParam

pub unsafe trait RValueParam<T>: Sized { }

A trait representing a parameter to a C++ function which is received by rvalue (i.e. by move).

Panics

The implementations of this trait which take a cxx::UniquePtr will panic if the pointer is NULL.

Safety

Implementers must guarantee that the pointer returned by get_ptr is of the correct size and alignment of T.

Object Safety

This trait is not object safe.