autocxx
Trait WithinBoxTrivial

pub trait WithinBoxTrivial: Sized + Unpin {
    // Required method
    fn within_box(self) -> Pin<Box<Self>>;
}

Emulates the WithinBox trait, but for trivial (plain old data) types. This allows such types to behave identically if a type is changed from generate! to generate_pod!.

(Ideally, this would be the exact same trait as WithinBox but this runs the risk of conflicting implementations. Negative trait bounds would solve this!)

Required Methods

fn within_box(self) -> Pin<Box<Self>>

Object Safety

This trait is not object safe.