autocxx
Module prelude

Imports which you’re likely to want to use.

Re-exports

pub use crate::as_copy;

pub use crate::as_mov;

pub use crate::as_new;

pub use crate::c_int;

pub use crate::c_long;

pub use crate::c_longlong;

pub use crate::c_short;

pub use crate::c_uchar;

pub use crate::c_uint;

pub use crate::c_ulong;

pub use crate::c_ulonglong;

pub use crate::c_ushort;

pub use crate::c_void;

pub use crate::AsCppMutRef;

pub use crate::AsCppRef;

pub use crate::CppMutRef;

pub use crate::CppPin;

pub use crate::CppRef;

pub use crate::CppUniquePtrPin;

pub use crate::PinMut;

pub use crate::RValueParam;

pub use crate::ValueParam;

pub use crate::WithinBox;

pub use crate::WithinBoxTrivial;

pub use crate::WithinUniquePtr;

pub use crate::WithinUniquePtrTrivial;

Macros

include_cpp

Include some C++ headers in your Rust project.

moveit

Performs an emplacement operation.

Structs

UniquePtr

Binding to C++ std::unique_ptr<T, std::default_delete<T>>.

Traits

Emplace

A pointer type that may be “emplaced” as a stable address which a New may be used to construct a value with.

New

An in-place constructor for a particular type.

Attribute Macros

cpp_semantics

Attribute which should never be encountered in real life. This is something which features in the Rust source code generated by autocxx-bindgen and passed to autocxx-engine, which should never normally be compiled by rustc before it undergoes further processing.