autocxx::subclass

Trait CppSubclassSelfOwnedDefault

pub trait CppSubclassSelfOwnedDefault<CppPeer: CppSubclassCppPeer>: CppSubclassSelfOwned<CppPeer> + Default {
    // Required method
    fn default_self_owned() -> Rc<RefCell<Self>>;
}

Provides default constructors for subclasses which implement Default and are self-owning.

Required Methods

fn default_self_owned() -> Rc<RefCell<Self>>

Create a self-owned instance of this subclass, initializing with default values. See CppSubclass for more details of the ownership models available.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.