autocxx
Module value_param

Traits

ValueParam

A trait representing a parameter to a C++ function which is received by value.

Functions

as_copy

Explicitly force a value parameter to be taken by copy.

as_mov

Explicitly force a value parameter to be taken using C++ move semantics.

as_new

Explicitly force a value parameter to be taken using any type of crate::moveit::new::New, i.e. a constructor.